% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_asdf.R
\name{as.data.frame.drift}
\alias{as.data.frame.drift}
\title{Coerce DRIFT-spectrum to a Data Frame}
\usage{
\method{as.data.frame}{drift}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{an object of \code{\link{S3-class}} \emph{drift}.}

\item{row.names}{NULL or a character vector giving the valid row names for the data frame.}

\item{optional}{logical. If TRUE "WN" prefix to wave numbers in column names is not added}

\item{...}{(not used)}
}
\value{
an object of \code{\link{S3-class}} \code{\link{data.frame}}
}
\description{
A method for the \code{\link{as.data.frame}} generics specified for
 objects of \code{\link{S3-class}} \emph{drift}. All meta tags are bound
 to the right of wave number columns.
}
\examples{
 s <- coal_drift()

 # default usage:
 dfs <- as.data.frame(s)
 stopifnot(is.data.frame(dfs))

 # use own names for cells:
 dfsn <- as.data.frame(s, c("measurement_1", "measurement_2"))
 stopifnot(is.data.frame(dfsn))

}
\seealso{
Other drift: \code{\link{cbind.drift}},
  \code{\link{coal_drift}}, \code{\link{diff.drift}},
  \code{\link{median.drift}}, \code{\link{plot.drift}},
  \code{\link{read.drift}}, \code{\link{slice}},
  \code{\link{write.drift}}
}
\concept{drift}
