% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_cbind.R
\name{cbind.drift}
\alias{cbind.drift}
\title{Combine DRIFT-spectra by columns}
\usage{
\method{cbind}{drift}(..., deparse.level = 1)
}
\arguments{
\item{...}{first, object(s) of \code{\link{S3-class}} \emph{drift}, then vectors or
matrices. These can be given as named arguments.}

\item{deparse.level}{integer controlling the construction of column names:
\describe{
  \item{0}{constructs no column names,}
  \item{1}{default, column names are combined as they are,}
  \item{2}{column names are enumerated.}
}}
}
\value{
An object of \code{\link{S3-class}} \emph{drift}.
}
\description{
Take a sequence of \code{\link{S3-class}} \emph{drift}, vector, matrix or
 data-frame arguments and combine by columns. Those columns are thus considered
 measurement cells. This is a \emph{S3}-method for generic function
 \code{\link{cbind}}.
}
\examples{
 # Light unit tests:
 x <- coal_drift("X")
 oc <- coal_drift("OC")
 s <- cbind(x, 4, oc, 2*rnorm(nrow(x)))
 stopifnot(class(s) == "drift")

}
\seealso{
Other drift: \code{\link{as.data.frame.drift}},
  \code{\link{coal_drift}}, \code{\link{diff.drift}},
  \code{\link{median.drift}}, \code{\link{plot.drift}},
  \code{\link{read.drift}}, \code{\link{slice}},
  \code{\link{write.drift}}
}
\concept{drift}
