% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_channel.R
\name{channel}
\alias{channel}
\title{Get the channel numbers of DRIFT-spectrum}
\usage{
channel(x, wn)
}
\arguments{
\item{x}{An object of \code{\link{S3-class}} \emph{drift}.}

\item{wn}{Numeric vector of wavenumbers for which channel numbers should be determined
in the DRIFT-spectrum}
}
\value{
Numeric vector of channel numbers which are interpolated values of indicies.
}
\description{
Get the channel numbers for appropriate wave numbers in the scale of the
 given object of \code{\link{S3-class}} \emph{drift}. The function is a
 simple backward approximation \emph{wavenumber - channel}.
}
\examples{
 s <- coal_drift()

 # Get channels of CO2 absorption:
 co2 <- channel(s, c(2268:2277, 2399:2412))
 print(co2)
}
\seealso{
Other drift: \code{\link{as.data.frame.drift}},
  \code{\link{cbind.drift}}, \code{\link{coal_drift}},
  \code{\link{diff.drift}}, \code{\link{median.drift}},
  \code{\link{plot.drift}}, \code{\link{read.drift}},
  \code{\link{slice}}, \code{\link{write.drift}}
}
\concept{drift}
