% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_coal_drift.R
\name{coal_drift}
\alias{coal_drift}
\title{DRIFT-spectrum of hard coal}
\usage{
coal_drift(x = c("X", "OC", "KO", "KC", "GXO", "GX"))
}
\arguments{
\item{x}{an abbreviation for coal rank.}
}
\value{
An object of \code{\link{S3-class}} \emph{drift}.
}
\description{
Get the example of DRIFT-spectrum for the given rank of hard coal from
 internal library as an object of \code{\link{S3-class}} \emph{drift}.
}
\details{
Latin character abbreviations of coal ranks used as values to argument
 \code{x} imitate those taken from \href{http://docs.cntd.ru/document/1200107843}{GOST 25543-2013}
 \emph{Brown coals, hard coals and anthracites. Classification according to genetic and technological parameters}
 as close as possible.

 The function is a primary constructor of objects of
 \code{\link{S3-class}} \emph{drift}. Other widely used constructor
 is \code{\link{read.drift}}.

 Objects of \code{\link{S3-class}} \emph{drift} can be indexed in the same
 manner as \code{\link{matrix}} objects. Though most of mathematical
 operation methos (see \code{\link{groupGeneric}}) inherite from
 \code{\link{matrix}} class some domain specific operations are rethinked.
}
\examples{
 s <- coal_drift()
 plot(s)

 # Indexing technique:
 #  - get first measurement cell in channels where CO2
 #    absorbtion is expected:
 s_co2 <- s[c(996:1000, 1064:1070), 1]
 plot(s_co2)  # liner segment is plotted for 1000-1064 1/cm range

 #  - get DRIFT-spectrum meta-data
 stopifnot(
   s[["ftirValidFileName"]] == "X",
   s[["gostAd"]] == 9.3
 )

 #  - set or reset new DRIFT-spectrum meta-data
 s[["ftirValidFileName"]] <- "BituminousBlendComponent"  # provide no spaces!
 plot(s)

 # Arithmetics with DRIFT-spectra:
 sgx <- -median(coal_drift("GX")) + 7.437333
 soc <- -median(coal_drift("OC")) + 7.004013
 wn <- 2920
 ds <- sgx - soc*as.double(slice(sgx, wn)/slice(soc, wn))
 plot(ds); abline(h = 0, v = wn)

}
\seealso{
Other drift: \code{\link{as.data.frame.drift}},
  \code{\link{cbind.drift}}, \code{\link{channel}},
  \code{\link{diff.drift}}, \code{\link{median.drift}},
  \code{\link{plot.drift}}, \code{\link{read.drift}},
  \code{\link{slice}}, \code{\link{write.drift}}
}
\concept{drift}
