% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_deriv.R
\name{deriv.drift}
\alias{deriv.drift}
\title{Differentiate or smooth DRIFT-spectrum}
\usage{
\method{deriv}{drift}(expr, ...)
}
\arguments{
\item{expr}{an object of \code{\link{S3-class}} \emph{drift}.}

\item{...}{other named parameters to \code{\link{smooth.spline}} and to
\code{\link{predict.smooth.spline}}, including the most
important ones:
\describe{
  \item{deriv}{integer order of derivative. First order derivative is the default value. Assign zero to perform smoothing which is the zero-order derivative}
  \item{spar}{smoothing parameter in (0,1] range. Default value 0.8 usualy is rather good value for DRIFT-spectra of hard coals}
  \item{df}{the desired equivalent number of degrees of freedom. Default value 12 usualy is rather good value for DRIFT-spectra of hard coals}
}}
}
\value{
An object of \code{\link{S3-class}} \emph{drift}.
}
\description{
Differentiate or smooth DRIFT-spectrum using spline fit model
}
\examples{
 s <- coal_drift()

 # first derivative:
 ds <- deriv(s)
 plot(ds)

 # simple smoothing of signal:
 ss <- deriv(s, deriv = 0)
 plot(ss)

 # second derivative:
 d2s <- deriv(s, deriv = 2)
 plot(d2s)

 # sequential application of differentiation is not equivalent to higher
 # order differentiation in place:
 range(deriv(deriv(s)) - deriv(s, deriv = 2))

}
\seealso{
Other drift: \code{\link{cbind.drift}},
  \code{\link{coal_drift}}, \code{\link{plot.drift}},
  \code{\link{read.drift}}, \code{\link{write.drift}}
}
\concept{drift}
