% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_diff.R
\name{diff.drift}
\alias{diff.drift}
\title{Differentiate or smooth DRIFT-spectrum}
\usage{
\method{diff}{drift}(x, n = 1L, ...)
}
\arguments{
\item{x}{an object of \code{\link{S3-class}} \emph{drift}.}

\item{n}{integer order of derivative. First order derivative is the default value. Assign zero to perform smoothing which is the zero-order derivative.}

\item{...}{other named parameters to \code{\link{smooth.spline}} including but not
limited to
\describe{
  \item{spar}{smoothing parameter in (0,1] range. Default value 0.8 usualy is rather good value for DRIFT-spectra of hard coals}
  \item{df}{the desired equivalent number of degrees of freedom. Default value 12 usualy is rather good value for DRIFT-spectra of hard coals.}
}}
}
\value{
An object of \code{\link{S3-class}} \emph{drift}.
}
\description{
Differentiate or smooth DRIFT-spectrum using spline fit model. This is
 an \emph{S3}-method for class \emph{drift}.
}
\examples{
 s <- coal_drift()

 # first derivative:
 ds <- diff(s)
 plot(ds)

 # simple smoothing of signal:
 ss <- diff(s, 0)
 plot(ss)

 # second derivative:
 d2s <- diff(s, n = 2)
 plot(d2s)

 # sequential application of differentiation is not equivalent to higher
 # order differentiation in place:
 range(diff(diff(s)) - diff(s, 2))
}
\seealso{
Other drift: \code{\link{as.data.frame.drift}},
  \code{\link{cbind.drift}}, \code{\link{channel}},
  \code{\link{coal_drift}}, \code{\link{median.drift}},
  \code{\link{plot.drift}}, \code{\link{read.drift}},
  \code{\link{slice}}, \code{\link{write.drift}}
}
\concept{drift}
