% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_median.R
\name{median.drift}
\alias{median.drift}
\title{Aggregate measurements of DRIFT-spectrum using median}
\usage{
\method{median}{drift}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{object of \code{\link{S3-class}} \emph{drift}}

\item{na.rm}{not used parameter for consistency with generic method}

\item{...}{potentially further arguments for methods; not used in the this method.}
}
\value{
An object of \code{\link{S3-class}} \emph{drift} that has only one
 measurement cell - the result of cell aggregation by \code{\link{median}}.
}
\description{
Given cells of measurements, make up the aggregated measurement with
 \code{\link{median}}. Due to non-normality in channel measurements in cells
 applying \code{\link{median}} instead of \code{\link{mean}} as an
 aggregation function seems relevant in
 \href{https://en.wikipedia.org/wiki/Diffuse_reflectance_infrared_fourier_transform_spectroscopy}{DRIFTS}.
 This is a \emph{S3}-method for generic function \code{\link{median}}.
}
\examples{

 # Light unit tests:
 x <- median(cbind(coal_drift(), coal_drift()*.92))
 plot(x)
 stopifnot(ncol(x) == 1)
}
\seealso{
Other drift: \code{\link{as.data.frame.drift}},
  \code{\link{cbind.drift}}, \code{\link{channel}},
  \code{\link{coal_drift}}, \code{\link{diff.drift}},
  \code{\link{plot.drift}}, \code{\link{read.drift}},
  \code{\link{slice}}, \code{\link{write.drift}}
}
\concept{drift}
