% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_nr.R
\name{nr}
\alias{nr}
\title{Convert DRIFT-spectrum to normalized radiance units}
\usage{
nr(x)
}
\arguments{
\item{x}{an object of \code{\link{S3-class}} \emph{drift}.}
}
\value{
an object of \code{\link{S3-class}} \emph{drift} exposed
  in \href{https://en.wikipedia.org/wiki/International_System_of_Units}{SI}-units
  of power: \eqn{W \cdot m^{-2} \cdot sr^{-1} \cdot \left ( {cm^{-1}} \right )^{-1}}{Wm^(-2)sr^(-1)(1/cm)^(-1)}.
}
\description{
Convert raw DRIFT-spectrum recorded in transmittance mode (i.e. reflectance)
 to normalized radiance units expressed in units of power as it calculated
 by \code{\link{planck}}.
}
\details{
Only raw DRIFT-spectra recorded into \emph{3709} channels within
  \emph{349.1157 - 7501.1654} \eqn{cm^{-1}} infrared range could be processed.
}
\examples{
gx <- median(nr(coal_drift("GX")))
oc <- median(nr(coal_drift("OC")))

vs <- gx - oc  # volatile substances
vs[["ftirValidFileName"]] <- "Volatile"
plot(slice(cbind(vs, oc, gx),,,3700))
abline(h = 0, col = "red")
}
\references{
\emph{Infrared Analysis of Particulates by FT-IR Emission/Transmission
 Spectroscopy}. Peter R. Solomon et al., \emph{Applied Spectroscopy},
 Volume: \emph{40} issue: \emph{6}, page(s): \emph{746-759}. DOI: \href{https://doi.org/10.1366/0003702864508421}{10.1366/0003702864508421}.
}
\concept{metric}
