% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_pinv.R
\name{pinv}
\alias{pinv}
\title{Calculate the Moore-Penrose generalized inverse of a matrix}
\usage{
pinv(x, tol = .Machine$double.eps^(2/3))
}
\arguments{
\item{x}{numeric matrix for which the \href{https://en.wikipedia.org/wiki/Generalized_inverse}{Moore-Penrose inverse} is required.}

\item{tol}{value of tolerance used for assuming an eigenvalue is zero.}
}
\value{
\code{y} - matrix that is pseudoinverse of matrix \code{x}
}
\description{
Calculate the generalized inverse \code{y} of a matrix \code{x}, also
  known as \href{https://en.wikipedia.org/wiki/Generalized_inverse}{Moore-Penrose inverse},
  using the singular value decomposition \code{svd()}.
}
\examples{
x <- matrix(c(7, 6, 4, 8, 10, 11, 12, 9, 3, 5, 1, 2), 3, 4)
b <- apply(x, 1, sum)  # 32 16 20 row sum
y <- pinv(x)
stopifnot(all.equal(drop(x \%*\% y \%*\% b), b))
}
\references{
Ben-Israel, A., and Th. N. E. Greville (2003). \emph{Generalized Inverses - Theory and Applications}. Springer-Verlag, New York.
  ISBN \emph{978-0-387-21634-8}.
}
\seealso{
Other math: 
\code{\link{planck}()}
}
\concept{math}
