% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_plot.R
\name{plot.drift}
\alias{plot.drift}
\title{Plot method for DRIFT-spectroscopy data}
\usage{
\method{plot}{drift}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of \code{\link{S3-class}} \emph{drift}.}

\item{y}{character vector of \code{\link{regex}}es that are searched against the list
of DRIFT-spectroscopy measurment attributes embedded in the object to
print them on plot.}

\item{...}{other parameters to \code{\link{plot.default}} and to \code{\link{lines}}.}
}
\value{
NULL, invisibly.
}
\description{
A method for the \code{\link{plot}} generics specified for
 objects of \code{\link{S3-class}} \emph{drift}. Method plots
 \href{https://en.wikipedia.org/wiki/Diffuse_reflectance_infrared_fourier_transform_spectroscopy}{DRIFTS}
 spectra in coordinates \emph{wavenumbers}-\emph{intensity} for middle
 \href{https://en.wikipedia.org/wiki/Infrared}{infrared} range.
}
\details{
For raw measurements in DRIFT-spectroscopy the \emph{intensity} represents the
proportion of diffuse \href{https://en.wikipedia.org/wiki/Reflectance}{reflectance}
radiated by specimen. Note that due to optic geometry in DRIFT-spectroscopy
the raw intensity in FTIR-spectrometer may assign \href{https://en.wikipedia.org/wiki/Transmittance}{transmittance}
units which actually is \emph{reflectance}.
}
\examples{
 s <- coal_drift()
 plot(s)
}
\seealso{
Other drift: \code{\link{as.data.frame.drift}},
  \code{\link{cbind.drift}}, \code{\link{coal_drift}},
  \code{\link{diff.drift}}, \code{\link{median.drift}},
  \code{\link{read.drift}}, \code{\link{slice}},
  \code{\link{write.drift}}
}
\concept{drift}
