% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_read.R
\name{read.drift}
\alias{read.drift}
\title{Read DRIFT data from Spectrotest Macro Measurements text file}
\usage{
read.drift(con)
}
\arguments{
\item{con}{a \code{\link{connection}} object or a character string}
}
\value{
An object of \code{\link{S3-class}} \emph{drift}
}
\description{
Read DRIFT data from \emph{Spectrotest Macro Measurements} text file
 or connection produced by \emph{FT-CONTROL} software of
 \emph{FTIR}-spectrometer
}
\details{
argument \code{con} has the same meaning as in \code{\link{readLines}}
}
\examples{
 #
 1 + 1
}
\concept{drift}
