% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_slice.R
\name{slice}
\alias{slice}
\title{Slice DRIFT-spectrum by wavenumbers}
\usage{
slice(x, k, left, right)
}
\arguments{
\item{x}{an object of \code{\link{S3-class}} \emph{drift}.}

\item{k}{a wavenumber, [\eqn{cm^{-1}}{1/cm}]. If given only one channel is extracted from
DRIFT-spectrum and further arguments are ignored.}

\item{left}{a wavenumber, [\eqn{cm^{-1}}{1/cm}]. If given indicates the lowest wavenumber
in wavenumber range.}

\item{right}{a wavenumber, [\eqn{cm^{-1}}{1/cm}]. If given indicates the highest wavenumber
in wavenumber range.}
}
\value{
An object of \code{\link{S3-class}} \emph{drift}.
}
\description{
Return a subset of a DRIFT-spectrum in the given range of wavenumbers.
}
\examples{
 s <- coal_drift()

 # extract one channel at 2000 1/cm:
 s1 <- slice(s, 2000)
 plot(s1)

 # make slice starting from 2000 1/cm:
 s2k <- slice(s, , 2000)
 plot(s2k)

 # make slice up to 1000 1/cm:
 s1k <- slice(s, , , 1000)
 plot(s1k)

 # subset in CO2 region:
 s2 <- slice(s, , 2399.44, 2411.02)
 plot(s2)

}
\seealso{
Other drift: 
\code{\link{as.data.frame.drift}()},
\code{\link{cbind.drift}()},
\code{\link{channel}()},
\code{\link{coal_drift}()},
\code{\link{diff.drift}()},
\code{\link{median.drift}()},
\code{\link{plot.drift}()},
\code{\link{read.drift}()},
\code{\link{write.drift}()}
}
\concept{drift}
