% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_write.R
\name{write.drift}
\alias{write.drift}
\title{Write DRIFT-spectrum object in Spectrotest format}
\usage{
write.drift(x, file = "")
}
\arguments{
\item{x}{an object of \code{\link{S3-class}} \emph{drift}.}

\item{file}{a \code{\link{connection}} object or a character string naming the file
to write to.}
}
\value{
NULL, invisibly.
}
\description{
Write object of \code{\link{S3-class}} \emph{drift} in Spectrotest
 format to text file or connection. Spectrotest format is a
 \href{https://en.wikipedia.org/wiki/Comma-separated_values}{CSV}-compatible
 text format with special \href{https://en.wikipedia.org/wiki/Tag_(metadata)}{meta-tags},
 so it could be easily opened by any
 \href{https://en.wikipedia.org/wiki/Spreadsheet}{spreadsheet software}.
}
\details{
Argument \code{file} has the same meaning as in \code{\link{write}}.
}
\examples{
 s <- coal_drift()

 # Write to text file:
 write.drift(s, "s.ir")

 # Write to connection:
 con <- xzfile("s.xz", open = "w")
 write.drift(s, con)
 close(con)
 rm(con)

 # Light unit tests:
 s0 <- read.drift("s.ir")
 stopifnot(all(s0 == s))
 unlink("s.ir")

 con <- xzfile("s.xz", open = "r")
 s0 <- read.drift(con)
 close(con)
 rm(con)
 stopifnot(all(s0 == s))
 unlink("s.xz")

}
\seealso{
Other drift: \code{\link{coal_drift}},
  \code{\link{plot.drift}}, \code{\link{read.drift}}
}
\concept{drift}
