% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_coal_drift.R
\name{coal_drift}
\alias{coal_drift}
\title{DRIFT-spectrum of hard coal}
\usage{
coal_drift(x = c("X", "OC", "KO", "KC", "GXO", "GX"))
}
\arguments{
\item{x}{an abbreviation for coal rank.}
}
\value{
An object of \code{\link{S3-class}} \emph{drift}.
}
\description{
Get the example of DRIFT-spectrum for the given rank of hard coal from
 internal library as an object of \code{\link{S3-class}} \emph{drift}.
}
\details{
Latin character abbreviations of coal ranks used as values to argument
 \code{x} imitate those taken from \href{http://docs.cntd.ru/document/1200107843}{GOST 25543-2013}
 \emph{Brown coals, hard coals and anthracites. Classification according to genetic and technological parameters}
 as close as possible.
}
\examples{
 s <- coal_drift()
 plot(s)

}
\seealso{
Other drift: \code{\link{cbind.drift}},
  \code{\link{diff.drift}}, \code{\link{plot.drift}},
  \code{\link{read.drift}}, \code{\link{write.drift}}
}
\concept{drift}
