% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_read.R
\name{read.drift}
\alias{read.drift}
\title{Read DRIFT-spectroscopy data}
\usage{
read.drift(con)
}
\arguments{
\item{con}{a \code{\link{connection}} object or a character string naming the file
to read.}
}
\value{
An object of \code{\link{S3-class}} \emph{drift}.
}
\description{
Read \href{https://en.wikipedia.org/wiki/Diffuse_reflectance_infrared_fourier_transform_spectroscopy}{DRIFTS}
 spectrum from \emph{Spectrotest} text file or
 connection produced by \emph{FT-CONTROL} software for
 \emph{FTIR}-spectrometers.
}
\details{
Argument \code{con} has the same meaning as in \code{\link{readLines}}.
}
\examples{
 s0 <- coal_drift()

 # Write to text file:
 write.drift(s0, "s0.ir")

 # Read from text file:
 s1 <- read.drift("s0.ir")

 unlink("s0.ir")

 # Light unit test:
 stopifnot(all(s0 == s1))
}
\seealso{
Other drift: \code{\link{cbind.drift}},
  \code{\link{coal_drift}}, \code{\link{diff.drift}},
  \code{\link{plot.drift}}, \code{\link{write.drift}}
}
\concept{drift}
